# Conflicts:
#	Translations/WinMerge/Slovak.po

+++ Externals/crystaledit/editlib/utils/cregexp.cpp
@@ -728,7 +728,7 @@ int RxMatch(RxNode *rx) {
                 if (*n->fChar != *rex) return 0;
                 if (memcmp(rex, n->fChar, n->fLen) != 0) return 0;
             } else {
-                for (size_t i = 0; i < n->fLen; i+= _tcsinc(&rex[i]) - &rex[i])
+                for (size_t i = 0; static_cast<short>(i) < n->fLen; i+= _tcsinc(&rex[i]) - &rex[i])
                     if (_totupper(mytcsnextc(&rex[i])) != _totupper(mytcsnextc(&n->fChar[i])))
                         return 0;
             }
